chore: print sizeof(struct jamfs_file_header)

+++ include/nos/fs/jamfs.h
@@ -7,11 +7,12 @@ struct jamfs_header {
   uint32_t nfiles;
 };
 
+// 76 bytes
 struct jamfs_file_header {
-  uint8_t magic;
+  uint32_t magic;
   char name[64];
   uint32_t offset;
   uint32_t length;
 };
 
-#endif  // !_NOS_FS_JAMFS_H
+#endif  // _NOS_FS_JAMFS_H
